# remove array specialization