@oscardssmith 2024-07-01
agreed. As I said, I still need to do some performance testing.