@oscardssmith 2024-07-02
Benchmarks:
```
using OrdinaryDiffEq
f(du, u, p, t) = du .= u
prob =ODEProblem(f, rand(1), (0.0, 1.0))

@btime solve(prob, Rosenbrock23())

# before
  12.360 μs (126 allocations: 9.62 KiB)
#after
  12.520 μs (126 allocations: 9.62 KiB)
  
@btime solve(prob, Vern7(lazy=false), abstol=1e-9, reltol=1e-9)
#before
  12.200 μs (220 allocations: 16.52 KiB)
#after
  12.330 μs (220 allocations: 16.52 KiB)
```

Seems to be well within the noise (especially considering that a single variable array is literally the worst case since  then any overhead would be magnified.