@ChrisRackauckas 2024-07-02
That's using `@btime` though. What's the compilation performance?