@oscardssmith 2024-07-02
and also, the other reason why this PR clearly is needed is that it found a bug in `Rosenbrock23` due to a previously untested path :)